+++ README.md
@@ -1,4 +1,4 @@
-# Templates to deploy FME Server in the cloud
+# IaC Templates to deploy FME Server in the cloud
 This repository is a collection of templates using different technologies to automatically provision cloud infrastructure and subsequently deploy FME Server on the provisioned infrastructure. These examples can be used as boilerplate code for custom deployments, make it easier to implement a reliable and reproducible deployment and allows to integrate the deployment of FME Server into existing CI/CD workflows.
 
 ## How to use this repository
@@ -8,3 +8,6 @@ The central piece of the distributed FME Server deployment are the VM images for
 
 ### Deployment templates
 Once the VM images are created they can be used as source images for the deployment. Alternatively any public source images provided by Safe Software can be used if available. The deployment templates implement a baseline of services needed for a distributed FME Server deployment but are not production ready out of the box and should be customized to each environment and use case.
+
+### Scaling Examples
+The scaling examples showcase a proof of concept to scale CPU-Usage (Dynamic) FME Server Engines based on queued jobs in Azure and AWS with python. These examples can be a starting point for an Azure Function and AWS Lambda implementation supporting different scaling scenarios.
