+++ ScalingExamples/scaleAzureVmss.py
@@ -1,35 +1,41 @@
-"""Scaling CPU-Usage (Dynamic) FME Server Engines (AWS)
+"""Scaling CPU-Usage (Dynamic) FME Server Engines (Azure)
 
 This script allows the user scale CPU-Usage (Dynamic) FME Server Engines with
-AWS ec2 Auto Scaling Groups (asg) based on the number of queued jobs for
+Azure Virtual Machine Scale Sets (VMSS) based on the number of queued jobs for
 a defined FME Server Queue.
 
 The constants are retrieved form environment variables:
 
-ASG_NAME        Name of the ec2 Auto Scaling Group
+RG_NAME         Name of the Azure Resource Group
+VMSS_NAME       Name of the Virtual Machnine Scale Set
 JOB_THRESHOLD   Threshold to scale the FME Server Engine either in or out 
 QUEUE           Name of the FME Server Queue
+SUBSCRIPTION_ID Azure Subscription ID
 FME_SERVER      FME Server URL
 FME_TOKEN       FME Server token
 
 This uses the FMEServerAPI.py as wrapper for the FME Server REST API
 Required modules:
 os
-boto3
+azure-identity
+azure-mgmt-compute
 """
 
 import os
-import boto3
+from azure.mgmt.compute import ComputeManagementClient
+from azure.identity import DefaultAzureCredential
 from FMEServerAPI import FMEServer
 
-ASG_NAME      = os.environ["ASG_NAME"]
-JOB_THRESHOLD = os.environ["JOB_THRESHOLD"]
-QUEUE         = os.environ["QUEUE"]
-FME_SERVER    = os.environ["FME_SERVER"]
-FME_TOKEN     = os.environ["FME_TOKEN"]
+RG_NAME         = os.environ["RG_NAME"]
+VMSS_NAME       = os.environ["VMSS_NAME"]
+JOB_THRESHOLD   = os.environ["JOB_THRESHOLD"]
+QUEUE           = os.environ["QUEUE"]
+SUBSCRIPTION_ID = os.environ["SUBSCRIPTION_ID"]
+FME_SERVER      = os.environ["FME_SERVER"]
+FME_TOKEN       = os.environ["FME_TOKEN"]
 
-aws_autoscaling_client = boto3.client('autoscaling')
-aws_ec2_client = boto3.resource('ec2')
+credential = DefaultAzureCredential()
+compute_client = ComputeManagementClient(credential, SUBSCRIPTION_ID)
 fme_client = FMEServer(FME_SERVER, FME_TOKEN)
 
 def activateScaleInProtection(instance_id: str) -> bool:
@@ -37,11 +43,13 @@ def activateScaleInProtection(instance_id: str) -> bool:
     Protects an instance from scale in operations
     """
     if instance_id:
-        aws_autoscaling_client.set_instance_protection(
-            AutoScalingGroupName=ASG_NAME, InstanceIds=[instance_id],
-            ProtectedFromScaleIn=True)
+        vm = compute_client.virtual_machine_scale_set_vms.get(
+            RG_NAME,VMSS_NAME, instance_id)
+        vm.protection_policy = {'protect_from_scale_in':True}
+        compute_client.virtual_machine_scale_set_vms.begin_update(
+            RG_NAME, VMSS_NAME, instance_id, vm)
     else:
-        return False
+        False
     return True
 
 def deactivateScaleInProtection(instance_id: str) -> bool:
@@ -49,11 +57,13 @@ def deactivateScaleInProtection(instance_id: str) -> bool:
     Remove protection from scale in
     """
     if instance_id:
-        aws_autoscaling_client.set_instance_protection(
-            AutoScalingGroupName=ASG_NAME, InstanceIds=[instance_id],
-             ProtectedFromScaleIn=False)
+        vm = compute_client.virtual_machine_scale_set_vms.get(
+            RG_NAME,VMSS_NAME, instance_id)
+        vm.protection_policy = {'protect_from_scale_in':False}
+        compute_client.virtual_machine_scale_set_vms.begin_update(
+            RG_NAME, VMSS_NAME, instance_id, vm)
     else:
-        return False
+        False
     return True
 
 def getInstanceId(hostname: str) -> str:
@@ -61,12 +71,12 @@ def getInstanceId(hostname: str) -> str:
     Retrieves an instance ID of the VMSS that matches the hostname of
     FME Server Engines 
     """
-    asg = aws_autoscaling_client.describe_auto_scaling_groups(
-        AutoScalingGroupNames=[ASG_NAME])
+    vms = compute_client.virtual_machine_scale_set_vms.list(
+        RG_NAME, VMSS_NAME)
     instance_id = ""
-    for i in asg["AutoScalingGroups"][0]["Instances"]:
-        if aws_ec2_client.Instance(i["InstanceId"]).private_ip_address == hostname:
-            instance_id = i["InstanceId"]     
+    for i in vms:
+        if i.os_profile.computer_name == hostname:
+            instance_id = i.id.split("/")[-1]      
     return instance_id
 
 def updateScaleInProtection() -> bool:
@@ -83,26 +93,24 @@ def updateScaleInProtection() -> bool:
             deactivateScaleInProtection(getInstanceId(i["hostname"]))
     return True
 
-def updateScaleSetCapacity(action: str, min_cap: int = 1) -> bool:
+def updateScaleSetCapacity(action: str, min_cap: int = 0) -> bool:
     """
     Update the scale in protection and scales the VMSS in or out based
     on the specified action
     """
-    updateScaleInProtection()
-    asg = aws_autoscaling_client.describe_auto_scaling_groups(
-        AutoScalingGroupNames=[ASG_NAME])
-    capacity = asg["AutoScalingGroups"][0]["DesiredCapacity"]
+    vmss = compute_client.virtual_machine_scale_sets.get(RG_NAME, VMSS_NAME)
     if action == "scaleOut":
-        aws_autoscaling_client.set_desired_capacity(
-            AutoScalingGroupName=ASG_NAME, DesiredCapacity=capacity+1)
+        vmss.sku.capacity += 1
     elif action == "scaleIn":
-        if capacity > min_cap:
-            aws_autoscaling_client.set_desired_capacity(
-                AutoScalingGroupName=ASG_NAME, DesiredCapacity=capacity-1)
+        updateScaleInProtection()
+        if vmss.sku.capacity > min_cap:
+            vmss.sku.capacity -= 1
         else:
            return False
     else:
         raise Exception ("Please specify scaleOut or scaleIn for action.")
+    compute_client.virtual_machine_scale_sets.begin_create_or_update(
+        RG_NAME, VMSS_NAME, vmss)
     return True
 
 def getQueuedJobs() -> int:
@@ -132,4 +140,3 @@ def scale() -> bool:
 
 if __name__ == "__main__":
     scale()
-
